Move markdown renderer component to shared
